Fully working file separation
